[RUN-3149] send large KeyIterationParams while iterating to accumulate keys, and shrink the PropertyMirrors array after (#216)
My hope is this won't result in much of a perf hit.  It seems (using the benchmark mode in https://github.com/replayio/backend/pull/9634) that it increases object preview durations by a constant factor, and things are still largely in the same ballpark as master.

Keep track of the passed-in params in the PropertyAccumulator, and in the case of `canOverflow` or `full` pass a MUCH larger params page size (currently 100k) down to the rest of the v8 machinery.  That way we're very likely to get enough keys for a given object that when we sort them we'll end up with the keys we expect.  This is coupled with early out'ing from the loop where we create PropertyMirrors, so the result sent back to the command JS is always min(num-properties, requested-size).

I did also find one spot where the key params weren't being passed, which looked like it resulted in getProperties on arrays returning all elements instead of just pageSize.  This makes a pretty huge difference in the benchmark durations for large arrays.